moveThoughtDown
